move ShieldedTransaction to shield/primitives
