+++ pallets/shield/Cargo.toml
@@ -34,6 +34,7 @@ sp-inherents.workspace = true
 
 # Subtensor primitives
 stp-io.workspace = true
+stp-shield.workspace = true
 
 [features]
 default = ["std"]
