+++ primitives/shield/Cargo.toml
@@ -7,6 +7,7 @@ edition.workspace = true
 
 [dependencies]
 codec.workspace = true
+scale-info.workspace = true
 sp-externalities.workspace = true
 sp-api.workspace = true
 sp-runtime.workspace = true
